Use gensym to uniquify locals instead of a counter
Makes analysis restartable, which is required when loop bindings and
recur expressions do not match in type and the loop body needs to be
analyzed. Fixes #199

This is a pretty big change that copies the tools.analyzer uniquify
pass into MAGIC along with the collect-closed-overs pass (because it
depends on uniquify). Semantically it should be fine but it might have
ramifications down the line.