Add bootstrapable load, compile and eval
The standard implementation of  clojure.core hardcodes invocations of
`clojure.lang.RT/load` and `clojure.lang.Compiler/eval` which makes it
impossible to change the compiler in use or the loading semantics.
This change includes a port of clojure.lang.RT/load to Clojure that
lives in core.clj and bottoms out in the dynamically bound functions
`*load-file-fn*` and `*compile-file-fn*` to load a script and compile
it to disk. similarly `eval` bottoms out in `*eval-form-fn*`.
Dynamically binding these functions allows the embedding system to
control these aspects of Clojure.

These functions use the dynamic var `*load-paths*` in place of the
environment variable that ClojureCLR uses.

To bootstrap from the standard compiler we do need to call
`clojure.lang.RT/LoadScript` from `*load-file-fn*` and
`clojure.lang.Compiler/eval` from `*eval-form-fn*` which this patch
does not do. To avoid a reference to the ClojureCLR compiler types
this can be done using reflection.